Call player classes
Improve the playscr so that it would call the player class and return the correct cells. Once this draws the new cells with the correct content, it flags the user to enter in coordinates to fire on and passes it to player.loadPlayerShot. Also added new logic to accept lowercase as well